+++ packages/Webkul/Paytm/src/Providers/EventServiceProvider.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace Webkul\Paytm\Providers;
+
+use Illuminate\Support\Facades\Event;
+use Illuminate\Support\ServiceProvider;
+
+class EventServiceProvider extends ServiceProvider
+{
+    /**
+     * Bootstrap services.
+     *
+     * @return void
+     */
+    public function boot()
+    {
+        Event::listen('checkout.order.save.after', 'Webkul\Paytm\Listeners\GenerateInvoice@handle');
+    }
+}