+++ packages/Webkul/Paytm/src/Resources/lang/de/app.php
@@ -1,0 +1,22 @@
+<?php
+
+return [
+	'checkout' => [
+		'onepage' => [
+			'payment' => [
+				'paytm' => [
+					'redirect' => 'Weiterleitung...',
+                    'redirecting' => 'Sie werden in wenigen Sekunden zu Paytm weitergeleitet.',
+                    'do-not-refresh' => 'Bitte aktualisieren oder schließen Sie dieses Fenster nicht.',
+                    'cart-empty' => 'Ihr Warenkorb ist leer.',
+                    'general-error' => 'Etwas ist schiefgelaufen. Bitte versuchen Sie es erneut.',
+                    'missing-cart-id' => 'Warenkorb-ID fehlt.',
+                    'cart-not-found' => 'Warenkorb nicht gefunden.',
+                    'checksum-failed' => 'Checksum-Validierung fehlgeschlagen.',
+                    'payment-failed' => 'Paytm-Zahlung fehlgeschlagen oder abgebrochen.',
+                    'payment-success' => 'Bestellung erfolgreich aufgegeben',
+                ],
+			],
+		],
+	],
+];