+++ packages/Webkul/Paytm/src/Resources/lang/en/app.php
@@ -1,0 +1,22 @@
+<?php
+
+return [
+	'checkout' => [
+		'onepage' => [
+			'payment' => [
+				'paytm' => [
+					'redirect' => 'Redirecting...',
+                    'redirecting' => 'You will be redirected to Paytm in a few seconds.',
+                    'do-not-refresh' => 'Please do not refresh or close this window.',
+                    'cart-empty' => 'Your cart is empty.',
+                    'general-error' => 'Something went wrong. Please try again.',
+                    'missing-cart-id' => 'Missing cart id.',
+                    'cart-not-found' => 'Cart not found.',
+                    'checksum-failed' => 'Checksum validation failed.',
+                    'payment-failed' => 'Paytm payment failed or cancelled. Please try again.',
+                    'payment-success' => 'Order placed successfully',
+                ],
+			],
+		],
+	],
+];