+++ packages/Webkul/Paytm/src/Resources/lang/fr/app.php
@@ -1,0 +1,22 @@
+<?php
+
+return [
+	'checkout' => [
+		'onepage' => [
+			'payment' => [
+				'paytm' => [
+					'redirect' => 'Redirection...',
+                    'redirecting' => 'Vous serez redirigé vers Paytm dans quelques secondes.',
+                    'do-not-refresh' => 'Veuillez ne pas actualiser ni fermer cette fenêtre.',
+                    'cart-empty' => 'Votre panier est vide.',
+                    'general-error' => 'Une erreur est survenue. Veuillez réessayer.',
+                    'missing-cart-id' => 'ID du panier manquant.',
+                    'cart-not-found' => 'Panier introuvable.',
+                    'checksum-failed' => 'Échec de la validation du checksum.',
+                    'payment-failed' => 'Le paiement Paytm a échoué ou a été annulé.',
+                    'payment-success' => 'Commande passée avec succès',
+                ],
+			],
+		],
+	],
+];