+++ packages/Webkul/Paytm/src/Resources/lang/he/app.php
@@ -1,0 +1,22 @@
+<?php
+
+return [
+	'checkout' => [
+		'onepage' => [
+			'payment' => [
+				'paytm' => [
+					'redirect' => 'מעביר...',
+                    'redirecting' => 'תועבר ל-Paytm בעוד מספר שניות.',
+                    'do-not-refresh' => 'נא לא לרענן או לסגור חלון זה.',
+                    'cart-empty' => 'העגלה שלך ריקה.',
+                    'general-error' => 'משהו השתבש. נסה שוב.',
+                    'missing-cart-id' => 'מזהה עגלה חסר.',
+                    'cart-not-found' => 'העגלה לא נמצאה.',
+                    'checksum-failed' => 'אימות Checksum נכשל.',
+                    'payment-failed' => 'התשלום ב-Paytm נכשל או בוטל.',
+                    'payment-success' => 'ההזמנה בוצעה בהצלחה',
+                ],
+			],
+		],
+	],
+];