+++ packages/Webkul/Paytm/src/Resources/lang/hi_IN/app.php
@@ -1,0 +1,22 @@
+<?php
+
+return [
+	'checkout' => [
+		'onepage' => [
+			'payment' => [
+				'paytm' => [
+					'redirect' => 'रीडायरेक्ट किया जा रहा है...',
+                    'redirecting' => 'कुछ ही सेकंड में आपको Paytm पर भेज दिया जाएगा।',
+                    'do-not-refresh' => 'कृपया इस पेज को रिफ्रेश या बंद न करें।',
+                    'cart-empty' => 'आपकी कार्ट खाली है।',
+                    'general-error' => 'कुछ गलत हो गया। कृपया पुनः प्रयास करें।',
+                    'missing-cart-id' => 'कार्ट आईडी नहीं मिली।',
+                    'cart-not-found' => 'कार्ट नहीं मिली।',
+                    'checksum-failed' => 'Checksum सत्यापन विफल रहा।',
+                    'payment-failed' => 'Paytm भुगतान विफल या रद्द हो गया। कृपया पुनः प्रयास करें।',
+                    'payment-success' => 'ऑर्डर सफलतापूर्वक किया गया',
+                ],
+			],
+		],
+	],
+];