+++ packages/Webkul/Paytm/src/Resources/lang/pl/app.php
@@ -1,0 +1,22 @@
+<?php
+
+return [
+	'checkout' => [
+		'onepage' => [
+			'payment' => [
+				'paytm' => [
+					'redirect' => 'Przekierowywanie...',
+                    'redirecting' => 'Za kilka sekund zostaniesz przekierowany do Paytm.',
+                    'do-not-refresh' => 'Nie odświeżaj ani nie zamykaj tego okna.',
+                    'cart-empty' => 'Twój koszyk jest pusty.',
+                    'general-error' => 'Coś poszło nie tak. Spróbuj ponownie.',
+                    'missing-cart-id' => 'Brak identyfikatora koszyka.',
+                    'cart-not-found' => 'Koszyk nie znaleziony.',
+                    'checksum-failed' => 'Weryfikacja checksum nie powiodła się.',
+                    'payment-failed' => 'Płatność Paytm nie powiodła się lub została anulowana.',
+                    'payment-success' => 'Zamówienie zostało pomyślnie złożone',
+                ],
+			],
+		],
+	],
+];