+++ packages/Webkul/Paytm/src/Resources/lang/zh_CN/app.php
@@ -1,0 +1,22 @@
+<?php
+
+return [
+	'checkout' => [
+		'onepage' => [
+			'payment' => [
+				'paytm' => [
+					'redirect' => '正在重定向...',
+                    'redirecting' => '您将在几秒钟后被重定向到 Paytm。',
+                    'do-not-refresh' => '请不要刷新或关闭此窗口。',
+                    'cart-empty' => '您的购物车是空的。',
+                    'general-error' => '出现错误，请重试。',
+                    'missing-cart-id' => '缺少购物车ID。',
+                    'cart-not-found' => '未找到购物车。',
+                    'checksum-failed' => '校验失败。',
+                    'payment-failed' => 'Paytm 支付失败或已取消。',
+                    'payment-success' => '订单已成功提交',
+                ],
+			],
+		],
+	],
+];