+++ packages/Webkul/Paytm/src/Routes/frount-routes.php
@@ -1,0 +1,19 @@
+<?php
+
+use Illuminate\Support\Facades\Route;
+use Webkul\Paytm\Http\Controllers\PaytmController;
+use App\Http\Middleware\VerifyCsrfToken;
+
+Route::group(['middleware' => ['web', 'theme', 'locale', 'currency']], function () {
+    Route::prefix('paytm')->controller(PaytmController::class)->group(function () {
+        Route::get('redirect', 'redirect')->name('paytm.redirect');
+
+        Route::get('cancel', 'cancel')->name('paytm.cancel');
+
+        Route::get('success', 'success')->name('paytm.success');
+    });
+});
+
+Route::post('paytm/callback', [PaytmController::class, 'callback'])
+    ->withoutMiddleware(VerifyCsrfToken::class)
+    ->name('paytm.callback');